Add estimate rpc
